+++ .travis/run.sh
@@ -1,0 +1,21 @@
+#!/bin/bash
+
+set -e
+set -x
+
+if [[ "$(uname -s)" == "Darwin" ]]; then
+    eval "$(pyenv init -)"
+    if [[ "${OPENSSL}" != "0.9.8" ]]; then
+        # so set our flags to use homebrew openssl
+        export ARCHFLAGS="-arch x86_64"
+        export LDFLAGS="-L/usr/local/opt/openssl/lib"
+        export CFLAGS="-I/usr/local/opt/openssl/include"
+        # The Travis OS X jobs are run for two versions
+        # of OpenSSL, but we only need to run the
+        # CommonCrypto backend tests once. Exclude
+        # CommonCrypto when we test against brew OpenSSL
+        export TOX_FLAGS="--backend=openssl"
+    fi
+fi
+source ~/.venv/bin/activate
+tox -e $TOX_ENV -- $TOX_FLAGS